requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
Werkzeug==3.1.3
pymysql==1.0.2
gunicorn==20.1.0
